Renaming agent entities: 1) Seer 2) Medic
